fix(db): store raw out-of-range menu ordinal on numeric put (MATCH-C, no clamp)
A numeric put to a menu(menuAlarmSevr)/menu(menuPini) severity/lifecycle
field clamped the ordinal to the valid range where C stores it verbatim.
C's putDoubleEnum is `*pfield = (epicsEnum16)*psrc` — a wrapping cast, not
a saturating clamp — so `caput REC.HSV 4` keeps 4 and `caput REC.HSV -1`
keeps 65535 (served signed as -1), and both derive C's garbage alarm
(STAT=HIGH/SEVR=INVALID). The port clamped at two sites for one invariant:
the typed AlarmSeverity/PiniMode carriers could not hold an out-of-range
ordinal, and `convert_to(Enum)` saturated a negative float to 0.

Structural fix — the carrier holds the raw 16-bit pattern:

- alarm/common_fields: AnalogAlarmConfig hhsv/hsv/lsv/llsv and CommonFields
  udfs/diss/pini become raw i16 (matching sel/dfanout's already-raw fields);
  alarm meaning is read with AlarmSeverity::from_u16(field as u16) and C's
  nonzero enable with `field != 0`.
- record_instance: menu_ordinal_raw() replaces parse_alarm_severity (stores
  the raw ordinal, no clamp); put/serve arms for HHSV/HSV/LSV/LLSV/UDFS/DISS/
  PINI carry it verbatim; evaluate_analog_alarm gates each level on `!= 0`
  and maps the ordinal through from_u16 so an out-of-range selector fires the
  level yet lands SEVR=INVALID/STAT=<level>, exactly as recGblResetAlarms.
- value.rs convert_to(Enum): the float→enum path now wraps via the i64 view
  (`(v as i64) as u16`) instead of saturating, matching C's (epicsEnum16)
  cast so a negative ordinal round-trips.

recgbl/bi/dfanout/mbbi/sel/processing/scan_index/iocsh adjusted for the raw
carriers. Oracle put phase: menu-ordinal family (427 cases across the seven
fields) now byte-identical to the compiled softIoc; 69 defects closed, 0 new.

+++ crates/epics-base-rs/tests/c_epics_parity.rs
@@ -165,10 +165,10 @@ fn ai_alarm_thresholds() {
         high: 70.0,
         low: 30.0,
         lolo: 10.0,
-        hhsv: AlarmSeverity::Major,
-        hsv: AlarmSeverity::Minor,
-        lsv: AlarmSeverity::Minor,
-        llsv: AlarmSeverity::Major,
+        hhsv: AlarmSeverity::Major as i16,
+        hsv: AlarmSeverity::Minor as i16,
+        lsv: AlarmSeverity::Minor as i16,
+        llsv: AlarmSeverity::Major as i16,
     });
 
     // Process to clear UDF alarm first
@@ -574,7 +574,7 @@ fn udf_alarm_on_uninit() {
     let inst = RecordInstance::new("TEST:udf".to_string(), rec);
 
     assert!(inst.common.udf != 0, "UDF should be true on new record");
-    assert_eq!(inst.common.udfs, AlarmSeverity::Invalid);
+    assert_eq!(inst.common.udfs, AlarmSeverity::Invalid as i16);
 }
 
 /// Multiple record types: field list completeness
@@ -1981,13 +1981,13 @@ async fn pini_flag() {
     // Set PINI
     if let Some(rec) = db.get_record("pini_rec").await {
         let mut inst = rec.write().await;
-        inst.common.pini = PiniMode::Yes;
+        inst.common.pini = PiniMode::Yes as i16;
     }
 
     // Verify PINI is set
     if let Some(rec) = db.get_record("pini_rec").await {
         let inst = rec.read().await;
-        assert_eq!(inst.common.pini, PiniMode::Yes, "PINI should be YES");
+        assert_eq!(inst.common.pini, PiniMode::Yes as i16, "PINI should be YES");
     }
     // `pini_records()` is the `piniProcess(menuPiniYES)` pass — it selects the
     // exact menu index, so only YES records are in it.
